requirements.txt
pandas
numpy
matplotlib
seaborn
pyyaml
jupyter
